requirements.txt
google-generativeai>=0.3.0
requests>=2.25.0
python-dotenv>=0.19.0 